Move importing and exporting of tasks through tars to a helper class
